pyproject.toml
[tool.poetry]
name = "sushi-grid"
version = "0.0.3"
description = "CLI tool for make sushi samples grid automatically"
packages = [
    { include = "app" }
]
authors = ["JoMingyu"]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Development Status :: 5 - Production/Stable",
    "Operating System :: MacOS",
    "Operating System :: Microsoft",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Typing :: Typed",
]
repository = "https://github.com/JoMingyu/sushi-grid"

[tool.poetry.scripts]
grid = "app:cli.handler"

[tool.poetry.dependencies]
python = ">= 3.9,< 4.0"
click = "^8.1.3"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
